refactor : Event Class 및  Event Observer BaseViewModel에서 분리
